swiper.el: Add a work-around for window-start not being current
* swiper.el (swiper--update-input-ivy): Update.

This results in double the window-height amount of lines being
heightlighted, instead of just window-height. But at least it doesn't
happen that some candidates within the current window aren't highlighted
since they're beyond the outdated window-start and window-end.

An alternative would be to use `redisplay' to update `window-start' and
`window-end', but that causes excessive blinking.

* ivy-test.el: Add a require.