New command ivy-partial-or-alt-done
In contrast to ivy-partial-or-done, this one always calls ivy-alt-done
if ivy-partial couldn't complete further.  This is more convenient to me
in the following situation:

With ivy-partial-or-done

  ~/some/d<tab>  => partial compl. to ~/some/di, 2 candidates: ding,  dir
  <tab>          => dired is opened for dir ~/some/ding/

With ivy-partial-or-alt-done

  ~/some/d<tab>  => partial compl. to ~/some/di, 2 candidates: ding,  dir
  <tab>          => ivy--cd into ~/some/ding/ and further completions
                    are possible

I'd suggest making ivy-partial-or-alt-done the default <tab> binding.
It seems to be more intuitive to me and is also more similar to ido.