ivy.el (ivy--insert-prompt): Use newlines instead of truncation
When the prompt string is longer than window-width, insert newlines
appropriately so that the whole prompt fits in the window without
scrolling.  Finally, when prompt+entered text would be larger than
window-width, reformat the prompt so that entered text starts on a
newline.  When completing file names, move the whole path to a new line
in that case.

Fixes #252