Switch to using ivy-exit-with-action
* ivy.el (ivy-exit-with-action): Add a missing quote.
(ivy--cd-maybe): Use `ivy-exit-with-action'.

* counsel.el (counsel-find-symbol):
(counsel--info-lookup-symbol):
(counsel-git-grep-query-replace): Use `ivy-exit-with-action'.

* swiper.el (swiper-query-replace):
(swiper-mc): Use `ivy-exit-with-action'.

The previous approach was overwriting the action list, so when
`ivy-resume' was called, only a single action was present.  The new
approach doesn't have this bug.

So now it's possible to e.g. `counsel-describe-function' -> "M-o d" ->
`ivy-resume' -> "M-o o" -> `ivy-resume' -> "M-o i".