When in prog-mode, "M-n" should wrap input with "\b"
When searching for occurrences of a symbol, it's useful to omit the
cases where the symbol is a part of a larger word.

"M-n" will insert the symbol, then if the mode is derived from
`prog-mode', wrap this symbol with \b - the regex notation for word
boundary.

This wrapping can be undone with `undo'.