ivy.el: hydra-style action prompt
Add a custom variable `ivy-action-display-style' to optionally toggle
the use of a hydra-style docstring as hint in the `ivy-read-action'
minibuffer prompt.

Add a variable `ivy--action-hints-lists' and a command
`ivy-set-action-hint' to customize this string globally or
per command. This gives some flexibility to arrange the layout of the
action prompt, which can be useful especially for commands with a large
number of actions.

Add a function `ivy-hydra--make-action-hint' which is called by
`ivy-read-action' to build the docstring when `ivy-action-display-style'
is set to 'hydra. This function is defined in ivy-hydra.el since it
depends on the hydra package, and is autoloaded in ivy.el.