Make the slash more magic during file name completion
* ivy.el (ivy--magic-file-slash): Extract from `ivy--exhibit'.
(ivy--exhibit): Update.

Update to the behavior: the slash ("/") will enter a directory even if
its name isn't completely typed out if either:

1. It's the only candidate.
2. The candidate index isn't 0, i.e. "C-n" has been typed at least once.
3. The input isn't "/".

The above rules still allow to keep the old behavior with "//" moving to
root and "/ssh:" opening tramp.

This is an experimental feature, please report if it breaks someone's
workflow.

Re #321.