ivy.el (ivy--reset-state): Add workaround for dired-do-copy
Example scenario: `dired-dwim-target' is t, and a file "foo" is copied
from directory-1 to directory-2. But directory-2 already has a file
"foobar".

Passed to ivy, :initial-input is "directory-2" and :preselect is
"directory-2/foo". Without the workaround, the :preselect will choose
"foobar" when "RET" is pressed. So the only correct way to make a file
copy and not a file overwrite would be "foo C-M-j" - much longer than
the indented "RET".

With the workaround, the preselect is disabled, resulting in "./" being
selected with "RET".