Extract ivy--flx-propertize from ivy--flx-sort
* ivy.el (ivy--flx-propertize): New defun.
(ivy--flx-sort): Don't propertize any more, only sort.
(ivy--format-minibuffer-line): Add propertize for fuzzy.

With the new change, the minibuffer propertize is done always for fuzzy.
Previously it would only be done if the number of cands was less than
200 in `ivy--flx-sort'.

Now, `flx-score' is called `ivy-height' times. That info is used to
propertize the visible cands.

Fixes #651