CD to home on `~/` not just `~`
Previously when using counsel-find-file, to find a file relative to
the home directory one needed to type `~$FILENAME`. This is counter to
lots of training/convention in using shells.

If you had typed `~/$FILENAME` it would be relative the root as `~`
would jump to home and then `/` would jump to root.

With this change it doesn't trigger the jump to home just on typing
the tilde, it waits for tilde-slash.