ivy.el: Allow toggling of sorting in ivy-read
Modify ivy-sort-functions-alist to allow a list of sorting functions to apply to
a collection. The car of this list is the current one. ivy-rotate-sort then
rotates this list through the different possibilities. I bound it to C-c
C-s. Here is a simple example to illustrate.

(setq ivy-sort-functions-alist
  '((read-file-name-internal .
    (ivy-sort-file-function-default string-lessp string-greaterp))
    (internal-complete-buffer . nil)
    (counsel-git-grep-function . nil)
    (Man-goto-section . nil)
    (org-refile . nil)
    (t . string-lessp)))

M-x find-file RET C-c C-s now switches from the default sorting to using
string-lessp. C-c C-s again switches to string-greaterp and so on.

ivy.el (ivy-sort-functions-alist): Allow list in cdr of each cell
ivy.el (ivy-rotate-sort): Add
ivy.el (ivy--sort-function): Add and use for
lookup in ivy-sort-functions-alist