Heed non-sort entries in ivy-sort-functions-alist
Currently, ivy--sort-function does not distinguish between
collection functions that are not members of
ivy-sort-functions-alist and ones whose associated sorting function
is nil, i.e. no sorting.

In ivy-completing-read, this meant that the lookup in
ivy-sort-functions-alist would always fall through to the default
sorting function, even for existing collection functions whose CDR
was explicitly set to nil.