counsel.el (counsel-org-goto--get-headlines): Fix point-min bug
outline-next-heading was always called once before starting the
processing of the headlines, which resulted in the first headline to
be excluded from the list if it was located at the very top of the
file i.e. at point-min.

The fixed function now checks if point is already on a heading and
only calls outline-next-heading if this is not the case before
starting the processing of the headlines.