Prevent infinite recursion when handlers call ivy
Previously, if a function in ivy-completing-read-handlers-alist called
ivy-completing-read, this would result in infinite recursion. However,
now if ivy-completing-read is called recursively without incrementing
"(minibuffer-depth)", it ignores all handlers on the second call,
breaking the recursion.

The result of all this is that functions in
ivy-completing-read-handlers-alist can now safely call
ivy-completing-read.