counsel.el (counsel-grep-like-occur): Add
* counsel.el (counsel-ag-occur): Update.
(counsel-rg-occur): Update.
(counsel-grep-occur): Update.

This fixes the (inconvenient and inconsistent) approach of using the cached
search result for `ivy-occur'. With the old approach, it was impossible to run a
search, save to *ivy-occur*, change and save the buffer(s), and re-run the
search from *ivy-occur* with "g". Now this ability is back.

Sample use:
    1. `counsel-rg' on a project for a variable to remove/refactor.
    2. save the search to *ivy-occur*.
    3. rewrite file1 and save it.
    4. press "g" in *ivy-occur* to refresh.
    5. rewrite file2 and save it.
    ...

Another issue that's fixed here is the removal of `rg' column limiting by width.
I use this setting:

    (setq counsel-rg-base-command
          "rg -i -M 120 --no-heading --line-number --color never %s .")

This means that lines longer than 120 will be clobbered by `rg' in order to
limit search output and speed it up.  But when saving to *ivy-occur*, I don't
want those lines to be clobbered any more.