ivy.el (ivy-completion-in-region): Use :initial-input and :unwind
Previously, Ivy's "prefixing" and "perfect match" index logic was
messed up for completion-in-region. The reason being is that the
initial prefix was never passed to Ivy: instead, the collection was
pre-filtered on the prefix and a "new" completion session didn't know
about it.

With this change:

- the initial prefix is passed to `ivy-read' as :initial-input,

- since the initial prefix is expected to remain on "C-g", a special
  :unwind lambda was added,

- the collection is still pre-filtered on the initial prefix, even if
you erase the initial input, the full candidates won't be restored.