ivy.el: Use text prop functions more accurately
Function remove-text-properties redundantly requires its third
argument to be a plist, so prefer the otherwise equivalent
remove-list-of-text-properties in order to a) reduce the chance of
mistakes and b) do what we mean.  The only apparent reason not to
use the latter is for XEmacs compatibility.

(ivy-read, ivy-cleanup-string, ivy--insert-prompt): Replace
remove-text-properties with remove-list-of-text-properties.