Improve swiper handling of invalid regexes
Now when splitting apart the input by space, we always regex escape
each part if it is an invalid regex.

For regex-ignore-order, this fixes a bug where in swiper you
couldn't hit RET to go to the line of the selected match if your input
was an invalid regex. ignore-order had logic to escape invalid
regexes, but swiper calls ivy--regex on the input, which previously
was not performing the same escaping.

For regex-plus, you now also get a literal search if you enter an
invalid regex (previously it silently failed to search). It seems like
this is an improvement in behavior, but perhaps it will annoy
regex-plus users.