ivy.el (ivy-immediate-done): Add condition on read-file-name
Added condition on read-file-name, it's all about read-file-name
compat after all.

Also, added comments around the rational for compat

It's assumed that any following changes of ivy--directory will be in
expanded form, so ivy-immediate-done will return dirs in expanded
form, just like before.

If that's not the case, one can consider to only return ivy--directory
unchanged if the user exits the minibuffer unchanged.