counsel.el: fix find invocation.
When evaluated, find * starts with shell expansion (which discards
dotfiles) which are then passed to find.

This commit fixes this call. The candidates are formatted differently
(candidates start with './') but the postprocess works the same.

Also, this commit changes the filter to discard only the content of
the .git subdir (and not the prefix ".git").

(counsel--list-files): new function.
(counsel--file-jump-function): new function.
(counsel--dir-jump-function): new function.
(counsel-file-jump): fix find invocation, port to windows.
(counsel-dir-jump): fix find invocation, port to windows.