Add find args customization for counsel-{file,dired}-jump
This is useful, because I use counsel in a big symlink farm, so the
default behaviour of "-type f" has to be repaced for me with "-type
f,l".  Other customization is e.g. inclusion of hidden files, as the
default is using * glob, that doesn't match hidden directories.

Fixes #1889