counsel.el (counsel-M-x): Optionally expand aliases
Added support via `counsel-M-x-transformer` to show the expansion of an
alias in a manner similar to how keybindings are shown. For example, if
`rb` is aliased to `revert-buffer`, and bound to `C-c r`, `M-x rb` will
show `rb (revert-buffer) (C-c r)`.

Currently the expansion is in the same font as the command, which looks
nice, but it would be easy to add support for a custom face.

The boolean `counsel-alias-expand` was introduced to enable or disable
this functionality, with it defaulting to non-nil.