swiper.el (swiper--update-input-ivy): Don't rely on window-start in terminal
`window-start' and `window-end' is not reliable when in a terminal. It
can be made reliable if a `redisplay' is performed, but then we get
annoying blinking.

So let's define `window-start' as the location of `window-height'
lines above `point'. We may add a few extra overlays here and there,
but the performance shouldn't suffer, since the area is roughly equal
to the window size anyway.

Fixes #1928