counsel.el (counsel-compile): Add
This provides the initial framework for counsel-compile. The building
of the various compile options is driven by a variable
counsel-compile-local-builds which provides a list of things that can
be used to build up the available compile commands. The list can
contain functions, strings or even just lists of options.

Currently just set to the default "make -k"

---
v2
  - fix setup to set up
  - accept atoms from funcall and listify
  - nconc the list

counsel.el (counsel-locate-git-root) more common functions

Introduce a "private" helper which can be used for general project
root finding and then use it for counsel-locate-git-root.

counsel.el (counsel-compile): add filtered compile history

To avoid the history polluting the current project we provide a
filtered history that weeds out any paths outside of the current
project directory. To do this we use metadata embedded in the
counsel-compile-history which should embedded the source and build
directories.

counsel.el (counsel-compile): add make completion

This adds a helper which can provide all the potential make
invocations for a given makefile. It differs slightly from the example
in helm-make in that it only considers PHONY targets which are
generally the top level "meta" targets of a makefile. This is still
quite a big list on some projects. For example QEMU provides 394 such
targets.

We could probably shorten the list somewhat by only considering PHONY
targets which we not in themselves prerequisites of other PHONY
targets but so far I've not implemented that as the additional
processing may be a performance issue.

The other differences from helm-make's approach is we do the
extract with plain shell pipes and sort the final result.

---
v2
  - defcustom for counsel-compile-make-args
  - defcustom for make pattern
  - use faces to visually separate blddir

counsel.el (counsel-compile): add build dir helper

This helper supports the common practice of using a build directory to
support multiple configurations of the build. This is often done to
avoid recompiling all build variants of a project when doing
development work.

---
v2
  - use recursive and cmd properties

counsel.el (counsel-compile--update-history): add compilation hook

We can't use ivy-read's history directly as we loose our propertized
strings. We also want to get useful information from M-x compile which
will do things like parsing for cd and other heuristics we'd rather
not complicate our lives with.

Repeatedly calling add-hook on each invocation is a little inelegant
but it's hardly going to show in the numbers.

doc/ivy.org: add counsel-compile to sample bindings

counsel.el (counsel-compile): add counsel-locate-project-dwim

Add a smarter project root function that tries a series of steps from
most emacsey to least emacsey.

Fixes #1941