counsel.el (add counsel-compile)
This provides the initial framework for counsel-compile. The building
of the various compile options is driven by a variable
counsel-compile-local-builds which provides a list of things that can
be used to build up the available compile commands. The list can
contain functions, strings or even just lists of options.

Currently just set to the default "make -k"

---
v2
  - fix setup to set up
  - accept atoms from funcall and listify
  - nconc the list