counsel.el (counsel-compile): add make completion
This adds a helper which can provide all the potential make
invocations for a given makefile. It differs slightly from the example
in helm-make in that it only considers PHONY targets which are
generally the top level "meta" targets of a makefile. This is still
quite a big list on some projects. For example QEMU provides 394 such
targets.

We could probably shorten the list somewhat by only considering PHONY
targets which we not in themselves prerequisites of other PHONY
targets but so far I've not implemented that as the additional
processing may be a performance issue.

The other differences from helm-make's approach is we do the
extract with plain shell pipes and sort the final result.

---
v2
  - defcustom for counsel-compile-make-args
  - defcustom for make pattern
  - use faces to visually separate blddir