counsel.el (counsel-compile--update-history): add compilation hook
We can't use ivy-read's history directly as we loose our propertized
strings. We also want to get useful information from M-x compile which
will do things like parsing for cd and other heuristics we'd rather
not complicate our lives with.

Repeatedly calling add-hook on each invocation is a little inelegant
but it's hardly going to show in the numbers.