counsel.el (counsel-compile): clean-up get-make-targets
The original code called a pipeline of shell functions to get the list
of PHONY targets. One of the consequence of this is we picked up a
bunch of noise if the make crapped out with an error. To avoid this we
bring all the text extraction into emacs with a temp buffer and only
process it if make is happy.

This is also a first step towards what we would need if we wanted to
use counsel--async-command to gather stuff and update the targets
dynamically.

---
v2
  - use bare form of (rx) for phony macro
  - tweak wording in comments
  - push/nconc/sort the results
  - revert de-hyphenating
  - let*->let

Fixes #1972