counsel.el (counsel-compile): new counsel-compile-env helper
This helper allows us to tweak counsel-compile-env. We don't use
separate actions. We either add the environment variable if it doesn't
already exist or remove it if it does.

There is scope for better validation and dealing with setting ARCH=foo
when we have ARCH=bar in the environment already.