counsel.el (counsel-compile): better handling for counsel--get-build-subdirs
It is possible the user might actually just have a single build
directory rather than a hierarchy. We can handle this case by simply
adding the probed directory to the list if we detect any non-directory
files in the directory. While we are at it we shouldn't be returning
any non-directories so filter those out.