ivy-test.el: Make friendlier
Do not pass arbitrary text as first argument to message.  Do not
clone remote Git repositories at toplevel.  Cloning repositories
shouldn't be needed at all for the purpose of accessing test data,
but even if it were, those commands belong in Makefile or similar,
not at the test suite's toplevel.

For discussion, see:
https://github.com/abo-abo/swiper/commit/7e735808e966d70340b5a2cef0430f913151d9b6
https://github.com/abo-abo/swiper/commit/fa87b977aecbc7ee5f3b77ddf94c6f040e17c348
https://github.com/abo-abo/swiper/commit/4171364ab3cdb21eabb45657a8d20bd85f3ecc9b

(counsel-find-file-with-dollars, counsel-find-file-with-dollars):
Use skip-unless.