swiper.el (swiper--isearch-function): Speedup for big files
Collect only point positions, not the lines themselves.

* ivy.el (ivy--remove-props):
(ivy--format):
(ivy--wnd-cands-to-str): Adapt to the case when the candidates aren't strings.

Perhaps the call to `ivy--format-minibuffer-line' can be moved to
`ivy-format-function'. This can result in a cleaner (and possibly more
flexible) code. But then every instance of `ivy-format-function' has
to be rewritten to apply `ivy--format-minibuffer-line'.

* ivy-test.el (swiper--isearch-format): Update.

* swiper.el (swiper-isearch-action): Adapt.
(swiper--line-at-point): Add.
(swiper--isearch-format): Update.