swiper.el (swiper--isearch-function): Change.
Having point before the last "o" in "dodo", as in "dod|o", should hop
to the beginning of the string upon entering "do" as the regexp.  This
is because when searching backward, the (match-end) of the match to
which we hop should not cross the starting point of the search.  When
changing the search string from "do" to "do?", we hop back to the
second occurrence of "do", since "do?" can match the second "d" taken
in isolation.