swiper.el: Fix overlay faces using regex-ignore-order
regex-ignore-order operates using a list of
regexps. swiper--update-input-ivy loops over each regexp, but
when swiper--add-properties is called we have forgotten which regexp
we are adding the overlay for, so it uses the first match face for all
matches. Now we pass along a "re-idx" argument that indicates the
index of the regexp being considered. swiper--add-properties gained
logic to choose the face corresponding to this index.

Fixes #2162