counsel.el (counsel-compile): favour project root over default-directory
The counsel-compile--current-build-dir variable is set if we ever
descend into a build directory. However sometimes the user might want
to build in the top level, for example QEMU's docker build targets are
independent of configured builds. In this case default-directory is a
poor choice if you are calling from a file somewhere deep in the
source tree, especially as we have a function to find the root.