ivy.el (ivy-occur-revert-buffer): Make more generic
* ivy.el (ivy-occur): Call the occur function with ivy--old-cands.
(ivy-occur-revert-buffer): Call the occur funciton with nil.
(ivy-switch-buffer-occur): In case nil was given, rescan the buffer
list for candidates.
(ivy--occur-default): Extract. The current null-cands logic is a bit
pointless, since if it was a list there's no way it changed; and if it
was "obarray", it's the old version of it, also unchanged.

Maybe rewrite functions that have the "obarray" as collection to
receive "'obarray" instead. So that we can revert using the new
version of "obarray".