swiper.el (swiper-query-replace): Add a more straightforward default
Example buffer contents with symlinks:

    .Xmodmap -> git/home/.Xmodmap
    .gitignore -> git/home/.gitignore
    .gitconfig -> git/home/.gitconfig

We search for "[^\n ]+\( -> \).*". This regex has 3 groups: the link,
->, and where it points to.

Note that we're using a relatively new `ivy--regex' feature, where an
explicit group "\( -> \)" acts as a splitter between the other two
implicit groups.

    (ivy--regex "[^\n ]+\\( -> \\).*")
    ;; => "\\([^\n ]+\\)\\( -> \\).*?\\(.*\\)"

Since there are 3 groups, a reasonable default for replacement is "\1
\2 \3".  In the end, we likely want to use "\3 \1"; the default serves
to remind us of the sytax.

The second default is: \,(concat \1 " " \2 " " \3).
It's a little more cumbersome, but it's Lisp! So we could use something like:

    \,(concat \1 " " \2 " " (expand-file-name \3))

Which I think may be powerful in some situations.