Create counsel-xdg-recentf to select recently used files in XDG-compliant programs.
Recently used files are located within the file "recently-used.xbel", located in
the directory specified in the environment variable "XDG_DATA_HOME" or in
"$HOME/.local/share". Emacs produces this directory by the function
"xdg-data-home".

The file is an XML file following the XBEL format, which is parsed using
"libxml-parse-xml-region" and traversed using functions provided in the Emacs
library "dom.el".

The standard from FreeDesktop.org (a.k.a. "fd.o" or "xdg") is described in the
following website:

https://www.freedesktop.org/wiki/Specifications/desktop-bookmark-spec/