Add XDG recent files to counsel-recentf.
- Include option to turn on or off as "counsel-recentf-include-xdg-list".
- Create function "counsel-recentf-candidates" to return recentf candidates. Use
  this function in counsel-recentf and counsel-buffer-or-recentf-candidates.
  - This function is where it is decided to include the XDG listed files.
- Because this feature uses the library "dom.el", it only works for Emacs
  > 25. Therefore, the XDG files list is gotten using a macro, to prevent the
  evaluation/compilation of functions that might not exist.
  - This new macro, "counsel--recentf-get-xdg-recent-files", uses "(require
    'dom)" after checking the version of Emacs.

Recently used files are located within the file "recently-used.xbel", located in
the directory specified in the environment variable "XDG_DATA_HOME" or in
"$HOME/.local/share". Emacs produces this directory by the function
"xdg-data-home".

The file is an XML file following the XBEL format, which is parsed using
"libxml-parse-xml-region" and traversed using functions provided in the Emacs
library "dom.el".

The standard from FreeDesktop.org (a.k.a. "fd.o" or "xdg") is described in the
following website:

https://www.freedesktop.org/wiki/Specifications/desktop-bookmark-spec/

Fixes #2403