counsel-kmacro: Create actions for cycling the ring and setting a saved counter value.
When cycling the keyboard macro ring, Emacs saves the current counter value with
the current macro definition, so that the value is restored when cycling back.

`counsel-kmacro' currently allows the user to run any defined macro with the
current counter value, and does not switch the counter value when running
different macros.  `counsel-kmacro-action-set-saved-starting-counter' allows the
user to change the counter value saved the the macro in `kmacro-ring'. This is
not the same as setting the current counter value.

`counsel-kmacro-action-cycle-ring-to-macro' allows the user to cycle the ring to
the selected macro, which uses `kmacro-cycle-ring-previous'.

Fixes #2502