ivy--compute-extra-actions does not dedup on "key"
When `actions` and `extra-actions` have the same "key" they
ivy--compute-extra-actions should dedupe the list of actions. Currently
it does not and it leads to some, in my opinion, unexpected behavior.

This change makes it so when there is a duplicate, we remove it from the
list of actions. Specifically, the duplicate from extra-actions is removed.