counsel.el (counsel-find-file-undo): Undo changing current directory
This means we can undo directory changes with "RET" or "DEL", while
using the standard `undo' binding.

Since we have this, we no-longer need the "DEL" preselect logic that
was making "RET" into and "undo" in that case.  Which is good, since
"DEL RET" is a valid workflow to select the parent directory in
`dired'.

Re #1424
Fixes #2593