counsel.el (counsel-compile): don't fall-back to default-dir (#2652)
We will never actually reach that case because a failing
counsel--compile-root errors out. The chances of being able to build
something in $HOME (or some random directory emacs was started in) is
quite small. If users want that behaviour then they should be
encouraged to add a function to the end of
counsel-compile-root-functions to return default-directory.

Notably none of the other callers of counsel--compile-root have a
fallback after.