counsel.el (counsel-compile): add support for compiling single file (#2652)
The default behaviour of M-x compile when we deal with a standalone
file outside of a project is to offer to build that one file. Support
that by adding a fall-back function at the end of the root finding functions
that catches that case and tweaks a buffer local version of
counsel-compile-local-builds that will return plain compiler
invocations.