Bump version to point release 0.13.3
Finish the manual sync between the upstream swiper.git monorepo and
the individual external branches in elpa.git that version 0.13.2
tried to reconcile, but that went only part of the way.

Once again, this is a bugfix, and does not constitute cutting a
full-blown release.  For discussion, see #2822 and the following
thread: https://lists.gnu.org/r/emacs-devel/2021-02/msg01935.html

* counsel.el:
* ivy-avy.el:
* ivy-hydra.el:
* ivy.el:
* swiper.el:
* doc/ivy.org: Bump version to point release 0.13.3.
* doc/ivy.texi: Regenerate.