Fix company-prefix and company-common when using counsel-company.
The company-common is always downcase, even when the prefix might not
be. This fix allows counsel-company to use company-common even when
the prefix has uppercase letters.

It also sets company-prefix to be equal to company-common, without
this, company-finish does not know that the prefix was expanded, and
does not correctly removes the prefix from the candidate, generating
duplicated characters.