Preserve idx text property in marked candidates
Both internal and external code expects the idx property to be on
the first character of the candidate.  Since ivy--mark actually
makes the mark a part of the candidate, it should also preserve any
text properties at the start of the candidate.

* ivy.el (ivy--mark): Copy any text properties from the start of the
candidate to the mark prefix.  Consistently use destructive
operations on ivy-marked-candidates.