Fix swiper preselect when initial-input specified
When swiper is called with an initial-input and the current point is
on a line that does not contain a match, the previous code would
return nil for preselect, which would position the first viable
candidate as the first candidate in the buffer.

This small change selects the first candidate at or following the
current point, which allows for a simple repeat-search function to
wrap a call to swiper.

* swiper.el (swiper): Fix preselect search condition (PR #2930).

Copyright-paperwork-exempt: yes