Support minibuffer-history-variable being disabled
Emacs 28 minibuffer history can be disabled with a value of t.

* ivy.el (ivy--update-history, ivy-reverse-i-search-kill): Leave
history variable alone if it is t.
(ivy--label-and-delete-dups): Avoid unnecessarily creating a hash
table for zero entries.
(ivy--reverse-i-search-symbol): Rename...
(ivy--reverse-i-search-history): ...to this more accurate name.
All uses changed.
(ivy-history-contents): Return nil if history variable is t.
Ensure ivy--reverse-i-search-history is set to a valid value.

Fixes #2923.